+++ Aras.VS.MethodPlugin.2019/Properties/AssemblyInfo.cs
@@ -1,0 +1,40 @@
+﻿//------------------------------------------------------------------------------
+// <copyright file="AssemblyInfo.cs" company="Aras Corporation">
+//     © 2017-2022 Aras Corporation. All rights reserved.
+// </copyright>
+//------------------------------------------------------------------------------
+
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following 
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("Aras Innovator Method Plugin")]
+[assembly: AssemblyDescription("Aras Innovator Method Plugin provides easy way to create and modify Aras Innovator server-side methods. Method code can be open or saved directly from/to an instance of Aras Innovator or from/to a local AML package. Aras Innovator Method Plugin for Visual Studio has also additional capabilities like generating template code to unit test the method; creating a new package directly in Aras Innovator prior to saving the method in it; etc. The Plugin provides extension for Visual Studio with all operations and project templates for different releases of Aras Innovator starting with Aras Innovator 11 SP8. Aras Corporation provides this plugin as is and will not provide any support or other related services related to this plugin.")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Aras Corporation")]
+[assembly: AssemblyProduct("Aras Innovator Method Plugin")]
+[assembly: AssemblyCopyright("© 2017-2022 Aras Corporation. All rights reserved.")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: InternalsVisibleTo("Aras.VS.MethodPlugin.Tests, PublicKey=002400000480000094000000060200000024000052534131000400000100010041a363e667fa2f8992977956a1168182e2902e09ae3ca7c42fd7b018498cc8b83c78df355364ad5be1e6d95ca8916b7cad3f568306d6a2be66f2ac3c6bd54390454868371e4b554f9d71108d006f43f50eb7f8cab190100fc01ba4dc8acd0fa42593410fbb7de452fe47cb852860d05865ff1d7c9c07deabc234f3ebc7d978d9")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// Version information for an assembly consists of the following four values:
+//
+//      Major Version
+//      Minor Version 
+//      Build Number
+//      Revision
+//
+// You can specify all the values or you can default the Build and Revision Numbers 
+// by using the '*' as shown below:
+// [assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion("1.1.0.0")]
+[assembly: AssemblyFileVersion("1.1.0.0")]