+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - [x] Add project template for Aras 14.0.6.
 - [x] Add project template for Aras 14.0.7.
 - [x] Add project template for Aras 14.0.8.
+- [x] Add support for Visual Studio 2022
 - [ ] Add Hotkey for plugin actions
 - [ ] Add Aras code anylizer to project templates
 
@@ -45,6 +46,9 @@ download in the
 
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.
+
+## 1.21
+- [x] Add support for Visual Studio 2022
 
 ## 1.20
 - [x] Added project template for Aras 14.0.8.